xen: Added support to build agent in xen builder
- Added platform detection setting to selectively turn off Wno-unused-local-typedefs
 for gcc 4.1.2
- Reverted changed in GetBuildInfo() as GCC 4.1.2 complained with warning,
  'type always evaluates to true'
- Fix in agent_param.cc, ~AgentParam() [Praveen]
- Fix in return data type CalculateBandwidth() - Fix in data type comparison in test_vgw.cc

+++ src/vnsw/agent/uve/vrouter_uve_entry.cc
@@ -747,7 +747,7 @@ uint8_t VrouterUveEntry::CalculateBandwitdh(uint64_t bytes, int speed_mbps,
     }
     uint64_t speed_bps = speed_mbps * 1024 * 1024;
     uint64_t bps = bits/diff_seconds;
-    return (double)bps/speed_bps * 100;
+    return bps/speed_bps * 100;
 }
 
 uint8_t VrouterUveEntry::GetBandwidthUsage
